chore: 서비스 이용안내 페이지 관련 명칭/URI 수정(service → guide)
