+++ src/routes/guide/GuidePrivacyPolicy.js
@@ -1,0 +1,37 @@
+import { Container } from "react-bootstrap";
+
+const GuidePrivacyPolicy = () => {
+  return (
+    <div>
+      <Container>
+        <p><h4><b>찰칵 개인정보 처리방침</b></h4></p>
+        <br></br>
+        <p>제1조 (개인정보의 처리 목적)</p>
+        <p>(주)찰칵은 다음의 목적을 위하여 개인정보를 처리합니다. 처리하고 있는 개인정보는 다음의 목적 이외의 용도로는 이용되지 않으며, 이용 목적이 변경될 시에는 별도의 동의를 받는 등 필요한 조치를 이행할 예정입니다.
+          회원 가입 및 관리 회원 가입 의사 확인, 회원 식별 및 인증, 회원 자격 유지 및 관리, 서비스 부정 이용 방지 등을 목적으로 개인정보를 처리합니다.
+          서비스 제공
+          사진 촬영 및 저장, 사진 찍은 장소 정보 저장 및 공유, 친구 추가 및 삭제, 친구와 함께 찍은 사진 태그 및 공유, 다른 회원들과 소통할 수 있는 커뮤니티 기능 등을 목적으로 개인정보를 처리합니다.
+          마케팅 및 광고
+          신규 서비스 개발 및 맞춤 서비스 제공, 인구통계학적 특성에 따른 서비스 제공 및 광고 게재 등을 목적으로 개인정보를 처리합니다.
+        </p>
+        <p>제2조 (개인정보의 처리 및 보유 기간)</p>
+        <p>
+          (주)찰칵은 법령에 따른 개인정보 보유·이용기간 또는 정보주체로부터 개인정보를 수집시에 동의 받은 개인정보 보유·이용기간 내에서 개인정보를 처리·보유합니다.
+          각각의 개인정보 처리 및 보유 기간은 다음과 같습니다.
+          회원 가입 및 관리: 회원 탈퇴 시까지
+          서비스 제공: 서비스 이용 계약 종료 시까지
+          마케팅 및 광고: 동의 철회 시까지
+          관계 법령 위반에 따른 수사·조사 등이 진행중인 경우에는 해당 수사·조사 종료 시까지
+          (주)찰칵의 이익을 위해 필요한 경우에는 별도로 보유할 수 있습니다.
+        </p>
+        <p>제3조 (개인정보 보호책임자)</p>
+        <p>
+          (주)찰칵은 개인정보 처리에 관한 업무를 총괄해서 책임지고, 개인정보 처리와 관련한 정보주체의 불만처리 및 피해구제 등을 위하여 개인정보 보호책임자를 지정하고 있습니다.
+          정보주체는 (주)찰칵의 서비스를 이용하시면서 발생한 모든 개인정보 보호 관련 문의, 불만처리, 피해구제 등에 관한 사항을 개인정보 보호책임자에게 문의하실 수 있습니다. (주)찰칵은 정보주체의 문의에 대해 지체 없이 답변 및 처리해드릴 것입니다.
+        </p>
+      </Container>
+    </div>
+  );
+};
+
+export default GuidePrivacyPolicy;
