+++ src/routes/guide/GuideTermsOfUse.js
@@ -1,0 +1,39 @@
+import { Container } from "react-bootstrap";
+
+const GuideTermsOfUse = () => {
+  return (
+    <div>
+      <Container>
+        <p><h4><b>찰칵 서비스 이용약관</b></h4></p>
+        <br></br>
+        <p>
+          제1조 (목적)
+          본 약관은 (주)찰칵이 운영하는 사진 공유 서비스(이하 "서비스")와 관련하여 (주)찰칵과 회원 간의 권리와 의무 및 책임사항을 규정함을 목적으로 합니다.
+        </p>
+        <p>
+          제2조 (정의)
+          본 약관에서 사용하는 용어의 정의는 다음과 같습니다.
+          "회원"이란 본 약관에 동의하고 (주)찰칵과 서비스 이용 계약을 체결하여 아이디(ID)를 부여받은 자를 말합니다.
+          "아이디(ID)"란 회원 식별과 서비스 이용을 위하여 회원이 설정하고 (주)찰칵이 승인한 문자 또는 숫자의 조합을 말합니다.
+          "비밀번호"란 회원의 정보 보호와 본인 확인을 위하여 회원이 설정하고 (주)찰칵이 승인한 문자 또는 숫자의 조합을 말합니다.
+          "게시물"이란 회원이 서비스에 게시한 사진, 글, 음성, 영상 등 모든 정보나 자료를 말합니다.
+          "공개 게시물"이란 모든 회원 및 비회원에게 열람 가능한 게시물을 말합니다.
+        </p>
+        <p>
+          제3조 (약관의 효력 및 변경)
+          본 약관은 서비스 화면에 게시하거나 기타의 방법으로 공지함으로써 효력을 발생합니다.
+          (주)찰칵은 필요한 경우 관련 법령에 위배되지 않는 범위 내에서 본 약관을 변경할 수 있습니다.
+          변경된 약관은 제1항과 같은 방법으로 공지함으로써 효력을 발생합니다.
+        </p>
+        <p>
+          제4조 (서비스의 내용 및 제공)
+          서비스는 다음과 같은 내용으로 구성됩니다.
+          사진 촬영 및 저장
+          사진 찍은 장소 정보 저장 및 공유
+        </p>
+      </Container>
+    </div>
+  );
+};
+
+export default GuideTermsOfUse;
